requirements.txt
cached-property==1.5.2
chevron==0.14.0
frozendict==2.4.4
genanki==0.13.1
lxml==5.2.2
python-docx==1.1.2
PyYAML==6.0.1
typing_extensions==4.11.0
